requirements.txt
Pillow
pytesseract
flask
gunicorn
flask-restful
